update 0.12.5v
- Update model and controller order info (Add done as boolean)
- Update controller mail (More mail)
- Update model article ( price is float now)
- Update controller user (use mail now)
- Update route (add validOrder route)
- Update test (Mail, Article, and order info)

+++ src/models/article.ts
@@ -38,7 +38,7 @@ export class Article extends Model {
         allowNull: false,
       },
       price: {
-        type: new DataTypes.INTEGER,
+        type: new DataTypes.FLOAT,
         defaultValue: 0,
       },
       picture: {
